+++ modules/graphics/src/cpp/GOWindow.cpp
@@ -422,31 +422,6 @@ void
 GOWindow::mousePressEventHandleRotate(QMouseEvent* e)
 {
     mousePositionOrigin = e->pos();
-    QRect plot_region(qtchild->geometry());
-    int x, y;
-    qtGetPosition(e, x, y);
-    GOAxis* h = findContainingAxis(goFig, remapX(x), remapY(y));
-    if (h) {
-        rotateCamera
-            = ((GOVectorProperty*)h->findProperty(GO_CAMERA_POSITION_PROPERTY_NAME_STR))->data();
-        rotateUp
-            = ((GOVectorProperty*)h->findProperty(GO_CAMERA_UP_VECTOR_PROPERTY_NAME_STR))->data();
-        rotateTarget
-            = ((GOVectorProperty*)h->findProperty(GO_CAMERA_TARGET_PROPERTY_NAME_STR))->data();
-        rotateForward = rotateTarget;
-        rotateForward[0] -= rotateCamera[0];
-        rotateForward[1] -= rotateCamera[1];
-        rotateForward[2] -= rotateCamera[2];
-        rotateSourceCameraDistance = sqrt(rotateForward[0] * rotateForward[0]
-            + rotateForward[1] * rotateForward[1] + rotateForward[2] * rotateForward[2]);
-        rotateForward[0] /= rotateSourceCameraDistance;
-        rotateForward[1] /= rotateSourceCameraDistance;
-        rotateForward[2] /= rotateSourceCameraDistance;
-        rotateRight = rotateUp;
-        rotateRight[0] = (rotateForward[1] * rotateUp[2]) - (rotateForward[2] * rotateUp[1]);
-        rotateRight[1] = (rotateForward[2] * rotateUp[0]) - (rotateForward[0] * rotateUp[2]);
-        rotateRight[2] = (rotateForward[0] * rotateUp[1]) - (rotateForward[1] * rotateUp[0]);
-    }
 }
 //=============================================================================
 void
@@ -492,44 +467,16 @@ GOWindow::mouseMoveEventHandleRotate(QMouseEvent* e)
 {
     QPoint dest(e->pos());
     QRect plot_region(qtchild->geometry());
-    GOAxis* h = findContainingAxis(
+    GOAxis* axis = findContainingAxis(
         goFig, remapX(mousePositionOrigin.x()), remapY(mousePositionOrigin.y()));
-    if (h) {
+    if (axis) {
         int x, y;
         qtGetPosition(e, x, y);
 
-        double az = -(x - mousePositionOrigin.x()) / 180.0 * M_PI;
-        double el = (y - mousePositionOrigin.y()) / 180.0 * M_PI;
-
-        std::vector<double> camera_position(rotateTarget);
-        camera_position[0] += rotateSourceCameraDistance
-            * (cos(el) * sin(az) * rotateRight[0] + -cos(el) * cos(az) * rotateForward[0]
-                + sin(el) * rotateUp[0]);
-        camera_position[1] += rotateSourceCameraDistance
-            * (cos(el) * sin(az) * rotateRight[1] + -cos(el) * cos(az) * rotateForward[1]
-                + sin(el) * rotateUp[1]);
-        camera_position[2] += rotateSourceCameraDistance
-            * (cos(el) * sin(az) * rotateRight[2] + -cos(el) * cos(az) * rotateForward[2]
-                + sin(el) * rotateUp[2]);
-        std::vector<double> camera_up(rotateTarget);
-        camera_up[0] = (cos(el + M_PI / 2.0) * sin(az) * rotateRight[0]
-            + -cos(el + M_PI / 2.0) * cos(az) * rotateForward[0]
-            + sin(el + M_PI / 2.0) * rotateUp[0]);
-        camera_up[1] = (cos(el + M_PI / 2.0) * sin(az) * rotateRight[1]
-            + -cos(el + M_PI / 2.0) * cos(az) * rotateForward[1]
-            + sin(el + M_PI / 2.0) * rotateUp[1]);
-        camera_up[2] = (cos(el + M_PI / 2.0) * sin(az) * rotateRight[2]
-            + -cos(el + M_PI / 2.0) * cos(az) * rotateForward[2]
-            + sin(el + M_PI / 2.0) * rotateUp[2]);
-        h->setThreeVectorDefault(GO_CAMERA_POSITION_PROPERTY_NAME_STR, camera_position[0],
-            camera_position[1], camera_position[2]);
-        h->setRestrictedStringDefault(
-            GO_CAMERA_POSITION_MODE_PROPERTY_NAME_STR, GO_PROPERTY_VALUE_MANUAL_STR);
-        h->setThreeVectorDefault(
-            GO_CAMERA_UP_VECTOR_PROPERTY_NAME_STR, camera_up[0], camera_up[1], camera_up[2]);
-        h->setRestrictedStringDefault(
-            GO_CAMERA_UP_VECTOR_MODE_PROPERTY_NAME_STR, GO_PROPERTY_VALUE_MANUAL_STR);
-        h->updateState();
+        axis->rotateCamera(
+            mousePositionOrigin.x(), e->pos().x(), mousePositionOrigin.y(), e->pos().y());
+        axis->updateState();
+        mousePositionOrigin = e->pos();
     }
 }
 //=============================================================================